refactor(backend-api): move the api route in separate files
Author: Rudy Savini - @SavGRY
